web_server_flask/pyproject.toml
[tool.poetry]
name = "web_server_flask"
version = "0.1.0"
description = ""
authors = ["Samuvel Antony <samuvel6826@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.21.0"
watchdog = "^4.0.1"
cmake = "^3.29.6"
opencv-python = "^4.10.0.84"
face-recognition = "^1.3.0"
pymongo = "^4.8.0"
firebase-admin = "^6.5.0"
flask = "^3.0"
flask-cors = "^4.0.1"
python-dotenv = "^1.0.1"
dlib = "^19.24.4"
face-recognition-models = "^0.3.0"
honcho = "^1.1.0"
flask-socketio = "^5.0"
python-socketio = "^5.11.3"
python-engineio = "^4.9.1"
websocket-client = "^1.8.0"
requests = "^2.32.3"
gevent = "^24.2.1"
colorlog = "^6.8.2"
opencv-contrib-python = "^4.10.0.84"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
setuptools = "^70.3.0"

[tool.poetry.group.dev.dependencies]
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
